+++ R/plot.R
@@ -10,7 +10,7 @@
 #' @importFrom stats setNames
 #' @importFrom ggplot2 ggplot aes geom_line geom_ribbon geom_point geom_hline
 #'  theme_bw theme scale_colour_manual scale_y_continuous sec_axis xlab ylab
-#'  ggplot_build element_blank
+#'  ggplot_build element_blank labs
 #' @importFrom rlang .data
 #'
 #' @return ggplot2 plot output
@@ -248,7 +248,7 @@ plot.proportion <- function(x, xlab = "Time", ...) {
     setNames(viridis::viridis(length(combos), end = 0.9), combos)
   }
 
-  ggplot2::ggplot(measure_df) +
+  p <- ggplot2::ggplot(measure_df) +
     ggplot2::geom_line(ggplot2::aes(
       x = .data$time,
       y = .data$y,
@@ -283,4 +283,19 @@ plot.proportion <- function(x, xlab = "Time", ...) {
     ggplot2::ylab("Modelled proportion of cases") +
     ggplot2::theme(legend.title = ggplot2::element_blank()) +
     ggplot2::xlab(xlab)
+
+  # Surface a non-default denominator: it has no other representation in the
+  # plot (unlike the numerator, which already appears in the pathogen legend).
+  all_pathogens <- unique(x$constructed_model$pathogen_names)
+  if (!is.null(x$denominator_combination) &&
+    !setequal(x$denominator_combination, all_pathogens)) {
+    p <- p + ggplot2::labs(
+      subtitle = paste(
+        "Denominator:",
+        paste(x$denominator_combination, collapse = ", ")
+      )
+    )
+  }
+
+  p
 }